Fix the disk IO throughput; quieten debugging.
I was being stupid with my waitcomplete loop.  Doing it until submitted == 0
is the sane thing to do. Duh.  With this, there's no artificial sleeps going on
in the disk code and I can hit what I think is the theoretical max on a single
disk.

Whilst here, clean up the debug printing behind DO_DEBUG