This is an initial cut at converting things over to use kqueue for notification.
It'd be nice to use lio_listio() to submit multiple IO requests at
once.

It'd also be nice if the kqueue filter for this would call aio_return
for us and stuff the results in said filter, so we could avoid having
to do a syscall for each individual IO that we're completing.